docs: update readme with build instructions, and remove toolchain path.
Now the source code is public, the toolchain path shouldn't be hardcoded
within the cmake file. This is now removed, and can be set by passing
-DTOOLCHAIN when building.

Some build instructions have been added to the readme. These are enough
to boot a Linux kernel up to the point of it hanging before starting
init.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>